The rdebug-call-queue system reworked, each entry is now a list of the command and options telling the system what it should do with the output. Currently, :tooltip and :info are supported, but more can easily be added (including support for lambda-expression, if needed).
The info buffer now only displays things if explicitly asked to do so.

Two proof-of-concept functions added, rdebug-pretty-print-to-buffer
and rdebug-pretty-print-region-to-buffer. Note that the user interface
might differ in the future (we don't want too many user-level
function.)

The info buffer is no longer treated as accumulative, this simplifies
the rdebug-cmd-XXX functions, but the user has to wait until the
command has terminated until the output is displayed in the info
buffer.


git-svn-id: svn://rubyforge.org/var/svn/ruby-debug/trunk@699 453b2852-3d18-0410-866e-d09c099698e4